+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release History
 
+## UNRELEASED
+
+- Add promotional site navigation
+- Add promotional home page
+- Add promotional site content page
+
+## 80.0 (2024-10-08)
+
+- Fix panels on category page
+- Update wycombe url for school hub
+
 ## 79.0 (2024-08-20)
 
 Compare: <https://git.torchbox.com/buckinghamshire-council/bc/compare/78.2...79.0>